Implement op_reply parser and serializer, msg_header serializer and add test cases

+++ src/utils.rs
@@ -60,8 +60,7 @@ pub fn pretty_dump(buffer: &[u8]) -> () {
     print!("\n");
 }
 
-/*
-pub fn u32_to_u8_array(x: u64) -> [u8; 4] {
+pub fn u32_to_u8_array(x: u32) -> [u8; 4] {
     let b3 = ((x >> 24) & 0xff) as u8;
     let b2 = ((x >> 16) & 0xff) as u8;
     let b1 = ((x >> 8) & 0xff) as u8;
@@ -69,4 +68,3 @@ pub fn u32_to_u8_array(x: u64) -> [u8; 4] {
 
     [b0, b1, b2, b3]
 }
-*/
